Split Fields over 3 separate modules
